Fix perf issues in Highlighted Jars.
Original implementation does too much work in the `RenderedWorld` event, both rescanning the current location and allocating a new list for the results. Also does unnecessary work in the draw itself in some cases, doing floating-point math per-draw and per-item to get an emote source rect that is essentially constant.

We can load the list of highlightable objects just once, on location change, and update it when objects are added to/removed from the location.

(It's possible to optimize this even further, e.g. by caching tile locations, layer depths, BC source rects and so on, but probably diminishing returns on those.)

Cleans up a few other code oddities like unnecessary creation of event lambdas and passing of `Helper` field.